Old DiagramTree ported to new change signaling mechanism
To simplify the removal of the old DiagramTree it got ported to
the new signaling mechanism via DiaApplication. Instead of
sprinkling a lot of diagram_tree_*() calls through the code the
DiagramTree object just listens on DiaApplication and Diagram
to react on the relevant changes.

Stuff like LayerDialog and 'recent files' should be ported, too.