Bug 568168 - allow to render "holes" with some renderers
A lot of Dia's bezier code has warnings like "only first BezPoint can be
a BEZ_MOVE_TO". This is basically a limitation of the bezier approximation
code originally used in the DiaGdkRenderer. It is currently used in the
renderer base class DiaRenderer, but there already are some renderers in
principle capable to handle move-to within bezier pathes just fine (cairo,
svg, wmf, ...).
Unfortuanetly also the XML format to serialize BezPoint has the hardcoded
assumption of only one MOVE_TO at the beginning of the data. Although this
is easily fixable by not using bezier(conn|shape)_(save|load) that would
result in broken forward compatibility (older versions wont be able any
longer to read bezier written with a newer version.
So for the moment the rendering of holes is restricted to 'intermediate'
rendering, e.g. the rendering of "Standard - Outline".

2009-01-18  Hans Breuer  <hans@breuer.org>

 * lib/bezier_conn.c(bezierconn_simple_draw) : unused function adapted to
 split a given bezier rendering at the move-to points into 'easier' pieces
 (bezierconn_save) : add g_warning() about not first BEZ_MOVE_TO
 * lib/beziershape.c(beziershape_simple_draw) : unused function adapted to
 render only up to the first move-to within a bezier path with holes
 (beziershape_save) : add g_warning() about not first BEZ_MOVE_TO
 * lib/dia_svg.c(dia_svg_parse_path) : instead of splitting into pieces at
 move-to within a path just return the whole thing (disabled)
 * lib/geometry.c(distance_bez_(line|shape)_point,) : instead of
 complaining about "BEZ_MOVE_TO found halfway ..." just calculate it
 * lib/diarenderer.[ch] : DiaRenderer::is_capable_to(RENDER_HOLES) to
 allow objects code to adapt to specific renderer capabilities
 * objects/standard/outline.c : render holes if the renderer used is
 capable to do so. Otherwise fall back to rendering 'inner' shapes with
 color_white
 * plug-ins/cairo/diacairo-renderer.c
   plug-ins/libart/dialibartrenderer.c : implement Renderer::is_capabe_to
 * plug-ins/wmf/wmf.cpp : make it capable to render holes