Bug 597599 - build error if neither libemf nor windows.h is available
BeginPath, PolyBezierTo, EndPath, FillPath and StrokePath are not available
with the compatibility layer wmf_gdi, ifdef the new _bezier() for that case.

Also renamed s/SAVE_EMF/DIRECT_WMF/ to better describe what it is going to
do. The incomplete metafile emulation layer was always shooting for WMF,
EMF would be a quite different beast.