[meta info] Implement "<ObjectMenu>/Follow Link..."
From the Object::meta::url it is now possible to follow links
either internally (just diagram opening) or via the shell. The
former is used when given meta string with either absolute or
relative path, but without some protocol prefix.
If there is some protocol prefix the string is interpreted as
external link an passed to gtk_show_uri (ShellExecute on win32)
otherwise the new internal function dia_file_open is used to
import and open the file.