Bug 304693 - Optionally embedded images
There is an extra property to select the embedding of images
into the diagram. If there are embedded images and a valid filename
the file is still preferred during load. But if it is not available
anymore the loading falls back to the embedded pixbuf.

If there is no filename at all the image embedding is enforced
and loading works with out complaints about missing files.