Align Connected: aligning at connected points
Aligning objects usually works on the objects bounding boxes.
To get straight lines connecting multiple objects in the same
direction it might be useful to align the connected points
instead. This will (slightly) move the objects - while staying
connected - to have the connection points on a straight line.

Currently implemented are horizontal and vertical alignment
based on the information given by the used connection points.
(Following commits will fix some of this info.)

move the objects