Comment on trouble with zlib1.dll
If zlib1.dll built with e.g. vc9 is loaded by Dia it may
silently exit the program with the call to gzdopen().
Probably this is one of Micro$oft security updates ;)
It looks like just another c-runtime mix-ups with fatal
outcome. The file number calculated by g_open() with msvcrt.dll
can not be valid with msvcr9.dll, but is this a reason to
exit the program?