[transform] edit angle and scales, fix move_handle
The Group transformation is supposed to consist of only rotation
and scaling. With it scaling in child-object coordinates and
rotating afterwards there is a three value matrix property editor.
Group::move_handle() is updated accordingly, so the transformation
use is dumbed down to something more understandable.
There is nothing enforcing these matrix limitations elsewhere,
so e.g. more complex matrices can still be imported or set by
other plug-ins. And there probably should not be such a limitation.