Bug 637611 - Mangled SVG import (incomplete line-to handling)
There is a serious limitation with Dia's standard bezier serialization.
It can only represent a single move-to at the beginning and following
bezier points are assumed to be curve-to.
There is a more complete implementation by data_add_bezpoint and
data_bezpoint, but using that would hamper forward compatibility.

So implemented a work-around in dia_svg_parse_path() which reproduces
line-to appearance also by curve-to control points.