Bug 576548 - [cairo] Highlighted text editing
Rather than full blown highlighting which spreads across the
whole renderer, this is just restricted to help text editing.
And instead of a single text color an alternate one gets used
if the contrast of the first one is just not strong enough.

This is not using any special drawing capabilities from cairo,
but could be implemented with the other interactive renderers,
too.