Bug 614278 - Only remove the active layer on load of initial diagram
An important invariant of DiagramData::active_layer is it being always
valid. As a result the initial object can not be created without a layer,
but that layer gets replaced when laoding a diagram.
Now with real load-into as in dropping another diagram/file into an
existing one this would destroy part of the diagram to preserve. So the
unusual side-effect of diagram_data_load() is restricted to the case
described above.