[autosave] non blocking auto save (disabled)
With embedded images the saving of diagram files finally got slow
enough to make the blocking of the GUI annoying. Now only the data
to be auto-saved is copied in the GUI thread. The in memory XML
creation (especially CPU sensitive for embedded images) and the
final diagram saving (might be slow with big files or significant
I/O latency) are done in a thread.

The code is currently disabled, because the error conditions
(message_* use) are not thread safe. It works flawlessly when
no error during save occurs.