New development-only plug-in: stress
This should not be installed anywhere, it's development only!
The only implemented function is 'eating' memory, to test
Dia under low memory conditions.
BEWARE: consecutive calls try to eat 9/10 of the available
memory. This will finally lead to hard program exit when
the next call to g_malloc() fails.
It is still useful to reproduce bugs like bug 643285 without
swapping, because the memory is just 'eaten' from the address
space, not from real memory. Win32 only for the moment.