Bug #591525 - independence of color and alpha editing
The alpha value is not any longer part of the color menu.
Instead there is an additional color button reflecting the
set alpha value and allowing to edit color and transparency.

For places where alpha is not useful the DiaColorSelection
can be instantiated without showing alpha at all.

This is basically solving my objections raised in
https://bugzilla.gnome.org/show_bug.cgi?id=591525#c16