Kludge for wrong clipboard clipping of EMF on Vista/Win7
Make dia/cairo scale of EMF OS version dependent. It works as
intended on WinXP, but on Vista and Win7 /64 only around 70% of
the "Copy Diagram" is within the automatic clipping region.

All experiments to make it work without kludge failed, I also
could not find an explanation for the behavior change on the
Internet.