build: fix wrong runtime usage of wmf plug-in
The recent cleanup of plug-ins/makefile.msc did break the wmf plug-in
by linking the static CRT. The result was bad-file-descriptor on final
fwrite(), so that's now detected as well. Finally there is some
preprocessor define to get a compile time error instead of producing
a less obvious runtime error.