Bug 576936 - leak of the RDP sheet
RDP is the French abbreviation for Petri net, but shape/sheet names
should be English. The main points of the shapes are of center. The
marquee looks strange - and why is it in green?
The RDP.sheet was never included with Dia so basically there was no
obvious way to create the respective objects.
It should cause no harm to actually remove the shapes as well.