Bug 488269 - empty undo actions when no properties are changed
This is not a complete fix, but some related behavior change.
The dict property was claiming a change even when not touched,
that's corrected now. But creating an empty change is necessary
to fulfill the contract of DiaObject::apply_props_from_dialog().

We could implement some silent ignoring of these empty changes,
but IMO it would be better to not enable the Apply/Ok at all,
if there is nothing to change.