dxf: respect specialties of SOLID and THICKNESS round-trip
The order of SOLID points differs from Dia's. Also it is the
only way AFAICT to fill something.
The line THICKNESS was inconsistent with dxf-import.c. For me
it does not work at all to exchange line-width with other DXF
importers. Line-weight (370) almost does, but only with one
importer so it's still disabled.
Finally there was a bug when converting to g_ascci_formatd()