Bug 598399 - PDF export landscape renderer regressions
GtkPrint was used for PDF pagination. But it has issues with
landscape due to modifying the page matrix and also showing
the result rotated in viewers - see bug 599401 in Gtk+.

With some DiaRenderer API extension (adding the update rectangle
to begin_render()) and the new data_render_paginated() the help
of GtkPrint is no longer necessary for PDF export.

(There is still something wrong with the clipping beginning on
the second page of the PDF, but that seems to be independent of
using the GtkPrint backend or cairo directly.)