[gtk-deprecated] obsolete TableDialog
With the newly developed ArrayProp editor there is no
need to have a custom made dialog anymore. The new one
is based completely on StdProps and only looses one
feature: constraints for Primary Key. But that's
requested to be removed by bug 668484 anyway.

The new dialog appearance could use some polishing
but that might come when turning UmlClassDialog ...