[substitute] move Diagram::selection_changed signal to base class
For the new object_substitute() facility an object within the active
selection can change. While object-add and object-remove already were
signals on DiagramData selection-changed was not. But it should
because DiagramData::selected is also managed in the base class.

When objects are added/removed the Diagram react on the signal to
update it's content, i.e. calls object_add_updates(obj, dia);