[substitute] Upgradeable Standard - ZigZagLine to Bezier
Accessible through the object's context menu this function substitutes
the _Zigzagline with a _Bezierline. The function creates a favorable
representation of the original _OrthConn data of the given object.

Also a small adaption of object_substitute() algorithm:
Bezierlines don't have just two connectable handles, but
every major handle is connectable. So if the 'input obj' is
at the end we should search subst from the end, too.