svg: layer import/export via named groups <g id="LayerName"></g>
Export: overwriting DiaRenderer::draw_layer() to wrap every layer's
objects into their own named group. This seems to be the common way to
transport layer information via SVG.

Import: Every top level item which is a group with a name/id we are converting
back to a layer. This is consistent with our SVG export and does
also work with layers from Sodipodi/Inkscape/iDraw/...