[substitute] new object 'Standard - Path'
Basically like "Standard - Beziergon" or "Standard - Bezierline" but
without the limit of move-to at start. To edit offer a conversion
to Bezier or other single standard elements: "Explode" (?).

To create allow to convert a selection of single elements
"Merge to Path". The default way of creation would be to import from
SVG or other path supporting importers.