svg: initial (hidden) shape import - no round-trip yet
If dropping a shape file to the diagram/toolbox the SVG filter
will get asked to load it. This get rid of 'Not a Dia File' from
the digram exporter as final fallback.

I wrote the functionality to diagnose the main point misplacement
using Steffen's "Shape Design" shapes, but import_shape_info() is
far from complete for round-trip - not sure where to store the extra
information ...