[restructuring] obsoletion of TextAttributes duplication
The StdProps system now optionally uses a second offset to get/set a
value on a Text object given by it's pointer. This way we can more
easily set a single text attribute via properties. Formerly doing so
from a scratch property was not possible due to missing values for
other attributes.
Almost all existing objects/ are converted to use the new scheme,
but the old way is still working at the moment.
The one object not converted is "UML - Object", it is doing some smart
things with TextAttributes which could not be replace mechanically.