[unit test] implement trivial DiaObject::draw() test
This is testing the new path renderer as well as the object's
draw method. Currently it is a just-doesn't-crash test, but with
a more reliable path renderer we might introduce checks e.g.
against the bounding box calculation of the object.

Also added dia_object_set_string() call to
_test_connectionpoint_consistency() to have some more realistic
object state to test.