pdf-import: experimental first version of the plug-in
Experimental PDF import plug-in based on http://poppler.freedesktop.org/
BEWARE: this is currently trying to interpret the whole file, there is
no way to select single pages for import. The text import could probably
be improved by using TextOutputDev internally, which reassembles the
original text chunks from the position of it's pieces. Currently they
are just converted to single text objects as they come.
The scaling of the input to Dia's coordinates system is just too simple.
More complex matrix operations lead to wrongly sized results. There are
even misplacements with Dia-generated render-test.pdf.

Depending on the PDF at hand this may bring the system to it knees ;(
(Especially the gradient simulation is to blame - also it seems to be
worthless without implementing path clipping first.