Use python-config where available
In the next Ubuntu release, python will be installed with multiarch support,
which breaks the build because assumptions are made about python library
locations which no longer hold true.

python-config has been supported in upstream Python since 2.6, so use
python-config to determine include file and library locations and names if it
is available. Otherwise fall back to the old behaviour.

This fixes the build in the current development release of Ubuntu, and is a
reasonably generic and cross-platform way to configure the build for Python.

Signed-off-by: Hans Breuer <hans@breuer.org>
(cherry picked from commit c66f9543c41008ccc2f093b744da6893c2cb91e9)