[warningectomy] implicit declaration of function 'filter_unregister_callback'
The non-win32 build was referencing a function not implemented at all.
It could not be called, due to being protected by plugin_can_unload(),
but that was just coincidence.
Now filter_unregister_callback() is defined and implemented, including
some error handling for callback's being invoked after unloading of
the providing plug-in.