svg: initial support for CSS style for svg-import
The style is a applied to all the basic objects as before, but now
it can come from a style block, with keys tag.class#id and all
variations. CSS styled files will now be much closer to the intent
after loading, but CSS styling is also quite expensive for two reasons:
 - we have to try all the variations of tag.class#id to find if
   there actually is some style (unavoidable)
 - we parse a referenced over and over again (because that was the
   easiest way to get it working)
See comment on _css_parse_style for major things missing.