svg: group transforms now pass the W3C SVG test suite
The initial attempt to translate matrix.x0 and matrix.y0 to
simple group offset was flawed. Now the given matrix is kept
as given and the groups position etc. are calculated with it.
The group' objects real coordinates are only moved without
a matrix, otherwise it's just a simple matrix change.

Also fix some comments/typos.