[transform] Group::move_handle() improvements
Group handles are only used for scaling of the group content.
The group position is supposed to be scale-invariant. With this
changes the handles behave as intended as long as there is no
rotation within the matrix.
In the rotation case the drawing might get skewed (acceptable
effect not specially handled elsewhere) and the size/position
of the group change (no idea how to properly avoid this yet).