svg: activate viewBox interpretation again on import
Now it is also applied on non-root svg elements defining a local
viewport. It is only used if we are _not_ importing shapes,
restricting otherwise was a mistake when shape import was added.
With the current shape set it was triggered very seldom , because
there are only 14 out of >700 shapes defining width, height and
viewBox.