Fix Beziergon behavior after restructuring
Adding and removing points (segments) to the Beziergon did not work as
before due to changes from 62dab7da68. Actually that commit did not take
into account, that bezierconn_closest_segment() returned 0 for the first
segment but beziershape_closest_segment() returned 1.
Now the segment index is zero-based as expected and the Beziergon code
use the appropriate offsets to deal with it.
Also a unit-test is specifically written for Add/Remove Segment to not
introduce such bugs with upcoming restructuring.