[clang-cl] Improved DIAVAR handling
While compiling Dia with clang-cl [1] there was a problem with multiple
definitions of variables declared with DIAVAR. Apparently the MS compiler
is more forgiving with inclusion of '__declspec(dllexport) Type name;'
in multiple translation units. clang-cl was producing colliding symbols
in different obj files. The implemented fix is treating DIAVAR and the
extern keyword as orthogonal concepts. So for a global variable the
declaration is 'extern DIAVAR SomeType name;' while the instantiation
is just 'DIAVAR SomeType name;'.
This also fixes a gcc complaint about an external variable being
initialized (in units.c).
Of course the best way to handle global variables is to get rid of them
altogether.

[1] http://blog.llvm.org/2013/11/the-clang-cl-fallback-mode.html