Bug 573261 - Avoid font kerning problems with pangocairo(fc)
There is a recurring bug with pangocairo related to kerning and font scaling.
See: https://bugzilla.gnome.org/buglist.cgi?quicksearch=341481+573261+700592
Rather than waiting for another fix let's try to implement the ultimate work
around. With Pango-1.32 and HarfBuzz the kludge in Pango is gone and apparently
substituted with a precision problem. If we now use huge fonts when talking
to Pango and downscale these with cairo it should work with all Pango versions.

Also add a sequence diagram used to understand the problem.