import: Implement DiaImportRenderer as a helper for further file import
This is finally the implementation of an old idea of mine: for file formats
which are basically serialized drawing commands it should be relative
straightforward to create the Dia objects via DiaRenderer interface. The WPG
format is in this category and the respective plug-in had initial code for
this idea for years. The following wpg-import.c is proof of concept and used
to debug this code.