[transform] Recalculate layer extents in layer_replace_object_with_list()
Currently this function is only used when un-grouping objects. Formerly
this was an operation never changing the extents because a group had
the same bounding box as it's joined objects. Now with groups with
transformation the objects position and size might change. But even
without the transformation this extra call should cause no harm except
for some extra CPU cycles.