cairo: do not compile cairo plugin if '--without-cairo' is defined.
Passing '--without-cairo' option to the build system disables
the inclusion of cairo header files but does not prevent the
code from building. As a result the compilation fails with
'implicit declaration of function' errors.

To fix this, modify the Makefile.am so that it does not compile
the plugin if '--without-cairo' is defined. Note that this
option must be passed by the user explicitly as the build system
will always detect libcairo on the system since GTK depends on it.

See: https://bugzilla.gnome.org/show_bug.cgi?id=729668
See: https://bugs.gentoo.org/show_bug.cgi?id=509636

Signed-off-by: Gokturk Yuksek <gokturk@binghamton.edu>