Prefer gtk_adjustment_configure() over new/set pattern
Previously Property::reset_widget() was creating and setting a new
adjustment object to reconfigure value and range. This might have been
the easiest way before gtk_adjustment_configure() was available with
GTK+ 2.14. But with the access pattern for PropEventHandler (next
commit) the recreation caused problems because of signals to already
destroyed objects. The easiest way to overcome that was the switch to
gtk_spin_button_get_adjustment() / gtk_adjustment_configure().