DiaRenderer::draw_arc() and fill_arc() now preserving direction info
The previous definition of Dia's draw/fill-arc was basically following
the definition coming from GDK as follows:

'''
width: the width of the bounding rectangle.
height: the height of the bounding rectangle.
angle1: the start angle of the arc, relative to the 3 o'clock position,
     counter-clockwise.
angle2: the end angle of the arc, relative to angle1

Draws an arc or a filled 'pie slice'. The arc is defined by the bounding
rectangle of the entire ellipse, and the start and end angles of the part
of the ellipse to be drawn.
'''

In contrast to GDK where angle2 is always greater than angle1 Dia's now
only uses the counter-clockwise rule requiring the backend code to do
it's own calculations.

The previous approach had shortcomings when backends are actually
interested in the direction of the arc. One such example is cairo,
which could use cairo_arc() and cairo_arc_negative().
If a more complex construct like a path for a rounded rectangle is done
the direction information becomes crucial to have a continuous path for
proper filling. The same issue can be observed with recently added
facilities like DiaPathRenderer and it's convert-to-path use.

Renderer core changes
 - gdk: counter-clockwise
 - import: direction preserving with big_arc and clockwise flags
 - path: XXX: still something wrong with rounded poly-line arcs
 - svg: now using directed arcs
 - transform: using path_build_arc() which migh be the XXX from above;)
Renderer plug-in changes:
 - cairo: now using directed arcs, also resurrecting draw_rounded_rect
 - cgm: counter-clockwise
 - drs: no change (now using directed arc rendering)
 - dxf: counter-clockwise
 - hpgl: counter-clockwise
 - libart: counter-clockwise (by line approximation as before)
 - metapost: already handled preserving arc direction
 - pgf: counter-clockwise
 - ps: counter-clockwise
 - pstricks: counter-clockwise
 - python: pass-through to plug-ins
   - diadisect.py: warn about angle difference being greater than 360°
   - diasvg.py: direction preserving
   - diastddia.py: direction preserving (removing FIXME from 2009)
   - export-renderer.py: no change, just pass-through
 - vdx: preserving direction (by control angle)
 - wmf: counter-clockwise
 - wpg: counter-clockwise
 - xfig: arc direction preserving