DiaRenderer: remove unused _rounded_rect_with_bezier() again
Now that main renderers are able to handle draw_rounded_rect() with the
help of base class implementation (draw_arc, draw_line) emulation this
copy is not necessary anymore (and unused since). Using it involved some
magic regarding the existence of draw_beziergon with the derived renderer
which did not look feel right either.