Bug 108770 - text_margin allows to attach text with some distance
Some while ago discussed on dia-list
https://mail.gnome.org/archives/dia-list/2014-March/msg00011.html
a text_margin property to allow Text alignment with some distance.
Also it is extending background fill by the margin. (With background
drawing on the distance_from method is now using the bounding box
e.g. for easier selection of text objects.
I'm not completely satisfied with the behavior of not connected text,
when the margin gets changed. It is currently moving the text although
just increasing the margin should probably just keep the text position
fixed and draw a bigger margin. OTOH this would make the object
behavior dependent on it's connection state.