vdx: adapt <ForeignData/> attributes and content
At least LibreOffice 4.2.4.2 is not happy with the padding
after base64 data, not even with the linebreak following. See:
https://bugs.freedesktop.org/show_bug.cgi?id=83275

The change in lib/prop_pixbuf.c to not put a linebreak into
the base64 data will have effect on other output as well,
but the tag directly following should not be a problem
anywhere.