Bug 735430 - vdx: Box rounding support for export and import
VDX 'Rounding' is basically the same as Dia corner_radius, so
draw_rounded_rect() was really easy to implement. While at it
remove all superfluous prototypes instead of adding a new one.

The import has to be a little more elaborated, because corner_radius
can only be handled with "Standard - Box", but VDX has no rectangle.
So we check the polygon for being rectangular and if so apply the
rounding.

Checked with ellipse.dia round-trip as well as with Visio Viewer 2010.

Also included some minor line style/caps tweaking.