path: make handle behavior equivalent to other objects
Shear and Rotate are now available via context menu.

move_handle() is supposed to be just moving that and related handles (e.g.
when N is moved NE and NW will move too). But 'opposite' handles are not
to be moved at all. So our transformation is invariant to the opposite
point of the path, not the object position.