[unit test] check bounding box calculation vs. path drawing
The use of DiaPathRenderer in _test_draw() is now hidden behind a call to
create_standard_path_from_object(). This way the 'bounds' of the drawing
can be measured indirectly.
Convert to path is supposed to create an equivalent representation which
should produce the same (or smaller) bounding box as the original object.

The exception list is currently pretty long, so there is definitely room
for improvement. With smaller default tolerance value the list would be
much longer. And finally it would be nice to not only check if the bounding
box is big enough, but also if is close to what's really needed. But this
initial version should help catching regressions already.