[transform] text rotation for "Standard - Text"
Rotations is done around the text handle position. This position is
only known to the object implementation. There is new render API
draw_rotated_text() taking the rotation center an angle, which is
much less invasive than making the underlying Text object aware of
angle and vertical alignment (and than not using it almost anywhere).
It is assumed that free text rotation will only be directly supported
by very few objects, so the burden should not be propagated to all
objects with text.
The initial version is should be working for almost all renderers
due to the default implementation of DiaRenderer::draw_rotated_text()
converting the text to path and rotating that.