build: Fix doc building
The installations were failing because doc/de/apa.html could not be found.
It was actually generated in doc/ directory.

As per xsltproc(1), the directories passed to `--output` flag need to end with /:

> This means, that e.g. -o directory will maybe not work, but -o directory/ will.

Otherwise, the program will consider the path a file and generate the outputs
relative to its parent.