sheet-editor: remove the object type indicator from the sheet editor
I'm not sure when it would be useful, since the user can't change it, and
it shouldn't make any visible difference. If needed, the plugin registering
the sheet could eventually be recorded; if it's custom.so, it's SVG, and
if not it says which plugin needs to be disabled/updated.