sheet-editor: remove sheets immediately in the sheet editor
The sheets would previously not be removed until an Apply, which kind-of
makes sense from a technical perspective --- they aren't gone yet, it's
just that their SheetMod::mod is SHEETMOD_MOD_DELETED --- but is
nonsensical from a user perspective. I hit Remove! Why is it still there?!

With this change, removed sheets are removed from the combobox immediately
upon the removal. This is done _after_ changing the combobox, as otherwise
the box would have an invalid iter as its current position.

It also removes some code from on_sheets_dialog_button_apply_clicked. This
code tries to reload the dialog from scratch, but ultimately causes a
use-after-free: the list store used by the combobox keeps references to
SheetMod objects that are freed in sheets_dialog_create.

Lastly, a new global variable 'store' was added to sheets_dialog.c.
Ideally, the dialog would be a GtkWidget itself instead of this structure,
however that's more work than one commit.