sheet-editor: fix the 'revert' button
The revert button featured the same use-after-free as the apply button,
fixed in 3ff6f508, where sheets_dialog_create invalidates the combobox
models.

This is fixed by manually resetting the list to correspond to the new
'original' field. This is because the SheetMod::sheet property is modified,
and so we have to reset it manually.