sheet-editor: ensure the Apply and Revert buttons are disabled unless needed
The infrastructure is actually in place to make sure that they are enabled
when needed, but that isn't visible until you poke around in the dialog.
Make sure they are _disabled_, so enabling them is visible.

Note that they don't handle cases where a change is done and then undone,
like 'Move ->' followed by 'Move <-'; this would be a bit harder, and I'm
not sure there's much benefit to it. Would be a neat idea in future,
though.