message: simplify MessageInternal
The prototype now only takes one va_list; va_copy is C99, I believe, and
so is probably available if any plugin needs it.

The implementations have been updated to just use g_strdup_vprintf. This
does mean a heap allocation, but (a) we shouldn't have that many log
messages, and (b) it avoids nearest_pow, which I'm trying to remove.