utils: drop use of dia_menuitem_key_quark
It's used twice in different contexts, which doesn't seem like a good idea,
and it also is in utils.h which I'm trying to delete.

Admittedly, using strings instead of quarks is a slight loss in terms of
stringly-typing, but I think it's probably worthwhile. In future, it might
make sense to avoid _set_data, but doing that with e.g. subclasses is
probably too much effort.