error: drop for now
This is only used in two places, neither of them in libdia, and both of
them can be replaced with GIOError.

Ideally, we wouldn't be making GIOErrors ourselves; however, it'd make
sense to eventually port this to GFile et al, which will make them.
Let's start using them now.

Eventually we'll probably want an error domain again, but we can handle
that once we actually know what we need to include