lib: only use diagtkrc on GTK+ 2
This isn't something that any other application does, and I'm not sure of
any good reasons to do it. GTK 4 also doesn't support it at all, and it
forces us to use filesystem paths where we might prefer GFiles in
future.

This also stops setting the default colour map on GTK+ 3: this function
doesn't seem to exist there, so it can just be removed now.