vdx: fix (most) compiler warnings in the VDX plugin
The odd variable names here overlap with the (unnamespaced)
structures in libdia. Change these to reduce the warnings.

Also, some of the assignments don't even do anything. The plugin
still seems to work, but it's probably somewhat buggy; bypassing
the warnings might not be the best idea, but it makes the build
log shorter.