colour: Stop initializing colormaps
color_init initialized the colormaps, which (I'm guessing) are useful
for low-colour support. However, they aren't available in GTK+ 3, and
it appears to work (at least on X.org 1.22.1.3, XWayland) without it.

This reduces the amount of work we do at startup slightly, and means that
on GTK+ 3 there's nothing interactive-specific to do in libdia_init. You
also don't have to initialize what is seemingly a regular data structure.