sheet-editor: use the GtkDialog::response signal in the sheet editor's Edit dialog
The edit dialog was previously kept in a global variable, and then free'd
when a button was pressed. However, if the dialog was closed some other
way (e.g. by pressing Esc), the variable would become invalid and mess
things up.

The thing is, the dialog was always a GtkDialog, and so listening to the
buttons directly is unnecessary. We can simply wait for a `response' signal
and figure everything out there. This means that we no longer need to keep
the variable around, and thus just need to make sure the single pointer is
free'd at the right time.