pixbuf: output images that are not pure black
Before this, including in Debian unstable's dia-0.97.3+git20220525-4,
exported images were always black. Fix this by not setting the Cairo
operation, and flushing the surface before future use.

In addition, the images were often not correct because they were off the
canvas. Set renderer->dia to avoid that. This also draws the background,
so we don't need that code, or indeed any cairo_t here.

The colour is no longer used, and the rect construction is easy enough to
just inline.