diasvgrenderer: Use the SVG 1.1 DTD
Fixes #390. We now claim support for SVG 1.1, which seems to be the most
accepted version for SVG 1.x (SVG 1.2 apparently didn't really go
anywhere.) Hopefully we actually conform!

The documentation for DiaSvgRenderer was also moved and updated to reflect
this change, since updating it to SVG 2.0 might be an ABI break. As part
of this, the GType work was moved into G_DEFINE_TYPE; we still do the
header manually because everyone loves reaching into the renderer.