build: enable more pointer and prototype warnings
-Wincompatible-pointer-types was useful for me trying to get Dia working
on GTK3, and doesn't seem like a bad idea in general.

-Wstrict-prototypes looks like it was planned to be enabled with GTK3, but
it seems to work fine with GTK2 after a handful of tweaks.

Several of the files affected are just missing (void) prototypes, but
a few relied on more implicit behaviour. PyDia was especially affected,
and now uses a bunch of casts to keep the prototypes around.

These are errors instead of warnings since they are basically never
correct, and we can silence them with casts, etc. if needed.