io: read/write xml documents with GIO streams
This is still blocking IO on main, so is far from our ideal end point,
but does at least give us the primary advantages of GIO, including that
we can mix-in coverters such as gzip {de,}compression.

This is especially handy as currently we rely on libxml2 ‘just doing
that’, which is especially problematic as 2.14, which we will now depend
on, no longer enabled gzip support by default.

Unlike diaXmlParseFile/xmlDiaParseFile which this seeks to replace we
make no attempt to handle bad ‘libxml1’ files as it's dubious that Dia
can meaningfully still support such files anyway.

See: https://gitlab.gnome.org/GNOME/dia/-/issues/569